requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
python-multipart
httpx
langchain
langchain-google-genai
google-generativeai
python-dotenv
sqlalchemy
newsapi-python
cryptography
huggingface_hub
dateparser==1.2.0
apscheduler==3.10.4
tweepy>=4.14.0
pytz